consolidates query and mutation tests into one file
